refactor: Controller 제거 및 리스트 렌더링을 RestaurantList에 이전
